+++ CHANGELOG.md
@@ -19,6 +19,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - added Russian language to myems-web
 - added Advanced Report Configs to myems-api
 - added Turkish language to myems-admin
+- added tbl_microgrids_users to myems_system_db in database
+- 
 ### Changed
 - changed fdd rule template in myems-admin
 - renamed tbl_microgrids_converters to tbl_microgrids_power_conversion_systems in database